scraping/pyproject.toml
[tool.poetry]
name = "scraping"
version = "0.1.0"
description = ""
authors = ["Joshua Higginbotham <jhigginbotham64@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Scrapy = "^2.4.0"
google-cloud-language = "^2.0.0"
google-cloud-storage = "^1.32.0"
stackoverflow-jobs = "^0.1.0"

[tool.poetry.dev-dependencies]
ipython = "^7.18.1"
notebook = "^6.1.4"
pandas = "^1.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
